backend/requirements-minimal.txt
# ==================================================================
# МИНИМАЛЬНЫЕ ЗАВИСИМОСТИ ДЛЯ ПРОЕКТА ONTHEGO
# ==================================================================
# Используйте этот файл для чистой установки проекта
# pip install -r requirements-minimal.txt
#
# Старый requirements.txt был переполнен лишними пакетами
# ==================================================================

# Django Framework
Django==5.2.8
django-environ==0.12.0

# Django REST Framework
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1

# CORS для работы с Flutter
django-cors-headers==4.9.0

# Фильтрация и поиск
django-filter==25.2

# API документация (Swagger/OpenAPI)
drf-spectacular==0.29.0

# Работа с изображениями
Pillow==11.0.0

# Валидация данных
pydantic==2.12.4
pydantic-core==2.41.5

# Работа с датами
python-dateutil==2.9.0.post0
pytz==2025.2

# ==================================================================
# ОПЦИОНАЛЬНЫЕ ЗАВИСИМОСТИ (закомментированы)
# ==================================================================
# Раскомментируйте при необходимости

# База данных PostgreSQL (если захотите перейти с SQLite)
# psycopg2-binary==2.9.11

# Redis и Celery для фоновых задач
# redis==7.0.1
# celery==5.4.0

# Debug Toolbar для разработки
# django-debug-toolbar==6.1.0

# Расширения Django
# django-extensions==4.1

# ==================================================================
# ВЕРСИИ PYTHON
# ==================================================================
# Python 3.11+ рекомендуется
# Проверено на Python 3.11.x
